requirements.txt
aiogram==3.0.0b7
python-dotenv
requests
yt-dlp
pydub
moviepy
yandex-music
ffmpeg-python
flask